rename es_epq_active to es_epqstate
es_epq_active was introduced in 27cc7cd, in the original discussion,
Andres mentioned he was *trying to signal that EPQ is currently
active from the POV of executor nodes* when choosing es_epq_active as
a field of EState. It seems to me we can rename es_epq_active to
es_epqstate and a non-null value would convey the same meaning.

Besides, a field name ending with "active" typically suggests a
boolean value.