Improve Parallel worker reporting in CREATE INDEX and REINDEX
Prior to this patch, only DEBUG1 logging displayed the number
of parallel workers planned for a CREATE INDEX or REINDEX
operation. Furthermore, the DEBUG1 did not show the number
of workers launched. This patch improves the situation by
allowing for a CREATE INDEX with a VERBOSE option which will
show both the number of parallel workers planned and the number
launched at INFO level. REINDEX already has a VERBOSE option and
it will also show the same parallel information as CREATE INDEX.

In the future, VERBOSE can be expanded to track progress for other
steps being performed by CREATE INDEX or REINDEX.

As parallel workers launched can only be determined in the Index AM's,
a verbose field was added to IndexInfo to instruct the Index AM to print
the logging information.