Redact user password on pg_stat_statements
Previously when using the pg_stat_statements extension and logging level is set
to DDL and an e.g CREATE USER or ALTER ROLE was executed, the entire SQL was
being logged into the pg_stat_statements view, including the user
password.

To replace hard coded values on SQL with $#, the node must have a
location field, so when JumbleQuery is executed, the location of these
nodes are stored on JumbleState.

This commit adds a location field on String type that is used to
represent the password, so that it can be redacted from logs. The
grammar for was also changed to fill the location value
{ CREATE|ALTER} {USER|ROLE|GROUP } identifier { [WITH] [ENCRYPTED]
PASSWORD 'value' }