Replace CaseTestExpr with a Param during planning.
For some time we've wanted to get rid of CaseTestExpr and replace it
with a less-spongily-defined node type, because we have had bugs
caused by the lack of a clear connection between the CaseTestExpr and
the expression node providing its value.  See for example commits
f0c7b789a and 14a158f9b.  This patch series implements a partial fix:
replace CaseTestExpr with a Param early in planning.

I tried to generate Params in parse analysis and not use CaseTestExpr
ever, but it turns out to be far too hard to keep the Params' IDs
distinct during rule rewriting, subquery pullup, etc.  It doesn't
seem really necessary either: in practice, the transformations that
risk confusion between different CaseTestExpr-using expressions don't
occur before eval_const_expressions.  So as long as we lock down the
association between a CaseExpr and its controlled nodes when
eval_const_expressions first sees the controlled nodes, we'll be okay.

It turns out to be a good idea to invent a new paramkind, here
called PARAM_EXPR, for this purpose.  I first intended to just use
PARAM_EXEC Params, but it's valuable to instead define PARAM_EXEC
as being for values passed between plan nodes while PARAM_EXPR is
for values passed around within a single expression.  In particular
this allows easily treating PARAM_EXPR Params as parallel-safe,
which they should be since we'll never break a single expression
across worker nodes.  The machinations around setParam signaling
should ignore these Params too.

This first patch installs the infrastructure for PARAM_EXPR Params
and causes CaseTestExpr nodes used for CASE to be replaced properly.
The various places that abuse CaseTestExpr for other purposes will
be dealt with in subsequent patches, with the end goal of removing
executor support for CaseTestExpr.