Remove contain_context_dependent_node function-inlining restriction.
Now that CaseExpr and ArrayCoerceExpr convert their CaseTestExpr
sub-nodes to Params, the sort of confusion that this test meant to
prevent is impossible, and indeed the search function cannot return
true anymore (except perhaps in estimation mode, where it won't
matter).  Let's remove the useless and optimization-defeating code.

Some of the regression tests added by commits f0c7b789a and 14a158f9b
are now dead code, in that the errors they intend to test for are
now impossible by design.  I left them there for now because the
fact that they still pass is good validation for this patch series.
I'm wondering whether to remove them in the final version though,
because it's hard to justify expending test cycles forever on dead
issues.