Convert nested assignments to use Params instead of CaseTestExprs.
The work in clauses.c is a little trickier than the previous steps,
because we can't just use find_casetestexpr to find the CaseTestExpr
to be replaced.  However, execExpr.c already had logic that knew
where to look for that, so we can just move that to clauses.c.

Also, there's a bit in ruleutils.c that I concluded could just
be dropped.  The code following that can deal with the situation
just as well, I think.

This is the last core-backend usage of CaseTestExpr, but we still
have to fix plpgsql ...