Remove plpgsql's abuse of CaseTestExpr.
We can use a PARAM_EXTERN Param almost as easily as a CaseTestExpr,
since the cast expression this is used in has no need to access
any other Params.

At this point, the executor's support for CaseTestExpr is dead
code, because such nodes will never get past the planner.
I'll leave removing that for a separate patch, though.