Optimize away useless Param set/load pairs.
The code generation patterns that the previous patches added to
execExpr.c will often give rise to this situation, so clean it up.

There might be more optimization work to be done here.
In particular, I'm not clear on whether it's worth replacing the
ExecJustApplyFuncToCase function removed in the previous patch.
I don't know what use-case that was intended to cover, so it's
hard to be sure what the new corresponding pattern is.