TEMP Get rid of TBMStatus
This is for simplification. It's useful to force everything to
the hash table for comparing its contents with the equivalent
radix tree logic.

Also, some one-page paths don't have coverage in regression tests,
so it would be nice not to have to worry about it.

TODO: Measure for regression
If creating a radix tree for a single page has too much overhead,
we can create its memory contexts lazily, but that's left for
future work. We can always put the one-page logic back if
necessary.