Earthdistance test should only list objects from the public schema
It is possible to run the test suite with additional extensions
installed to make sure that everything works with them. If the
additional extension is installed in a custom schema it doesn't change
any of the normal test outputs.

Except for the eartdistance test, which lists all objects from every
schema.

This commit modifies the testcase so that it only selects objects from
the public schema to avoid this. This was already the case for part of
the test, these changes that unify everything in it.