pg_basebackup via pg_createsubscriber
This patch adds support for initializing a standby node using 'pg_basebackup'
through the 'pg_createsubscriber' tool. A new '--create-standby' option is
introduced, which automates the process of performing a base backup from the
publisher, and configuring the standby node for streaming replication.

To support this functionality, the patch introduces helper functions for
invoking 'pg_basebackup', and configuring the standby server.
It also adds support for managing external binary paths for 'pg_basebackup' and
'psql', making the tool more flexible in varied environments.

By integrating 'pg_basebackup', this patch enhances the usability of
'pg_createsubscriber' by enabling a streamlined, end-to-end setup of
subscriber nodes.