Generate the WaitClassTable[]
The WaitClassTable[] array that will be used as a lookup table by the wait events
statistics (implemented in a following commit).

The array is indexed by classId (derived from the PG_WAIT_* constants), handles
gaps in the class ID numbering and provides metadata for wait events.

A new struct (WaitClassTableEntry) is used to store the wait events metadata.

This new array is generated in generate-wait_event_types.pl, so that:

- it now needs lwlocklist.h and wait_classes.h as input parameters
- WAIT_EVENT_CLASS_MASK and WAIT_EVENT_ID_MASK have been moved away from
wait_event.c

In passing it adds severals new macros that will be used by a following commit
implementing the wait events statistics.

This commit does not generate any new files but adds some code in
wait_event_types.h and pgstat_wait_event.c.