Filter transactions that need not be published.
This patch set aims to filter transactions at the decode stage rather than
streaming time, which allows the system to skip processing transactions that do
not contain tables included in the logical replication walsender's publication
list. As a result, this can prevent the serialization of unnecessary records to
disk during non-streaming mode, especially in large transactions, and also
reduce memory and CPU usage in streaming mode when many changes can be filtered
out.

A hash table of relation file locators is implemented to cache whether a
relation is filterable or not. This ensures that we only need to retrieve
relation and publication information from the catalog when a cache entry is
invalid, avoiding the overhead of starting a transaction for each record.

Filtering is temporarily suspended for a sequence of changes (set to 100
changes) when an unfilterable change is encountered. This strategy minimizes
the overhead of hash searching for every record, which is beneficial when the
majority of tables in an instance are published and thus unfilterable. The
threshold of 100 was determined to be the optimal balance based on performance
tests.

Additionally, transactions containing WAL records (INTERNAL_SNAPSHOT,
COMMAND_ID, or INVALIDATION) that modify the historical snapshot
constructed during logical decoding are deemed unfilterable. This
is necessary because constructing a correct historical snapshot
for searching publication information requires processing these WAL records.

Note that this patch filters changes only for system catalog relations,
non-logically logged relations, or temporary heaps and sequences. A subsequent
patch will introduce a new output plugin in pgoutput, which will further filter
changes for relations not included in publications.