Fix a race condition in ConditionVariableTimedSleep()
A process could exit ConditionVariableTimedSleep() after detecting a state
change (cv != cv_sleep_target) without being removed from the list.

When ConditionVariableTimedSleep() detects that the process is no longer
in the wait list (could be signaled by ConditionVariableSignal() or
ConditionVariableBroadcast()), it re-adds itself to ensure that we don't miss
any additional wakeup.

The issue is that it could leave ConditionVariableTimedSleep() with cv_sleep_target
set to NULL and the process re-added in the wait list.

If cv_sleep_target is set to NULL then subsequent calls to ConditionVariableCancelSleep()
will return early without removing the process from any list.

This could to an assertion failure when the process later tries to wait on a
condition variable, as ConditionVariablePrepareToSleep() skips cleanup
when cv_sleep_target is NULL, resulting in a process trying to be in two wait
lists simultaneously.

The fix re-assigns cv_sleep_target to cv and then ensures that cv_sleep_target
accurately describes which condition variable weâre prepared to wait on.