Add support for runtime parameters in injection points
The macros INJECTION_POINT() and INJECTION_POINT_CACHED() are extended
with an optional argument that can be passed down to the callback
attached, giving to callbacks the possibility to manipulate a state
given by the caller.

93bc3d75d8e1 (test_aio) and da7226993fd4 (core AIO infrastructure) have
been relying on set of workarounds where pgaio_inj_cur_handle is used as
runtime argument given to the injection point callbacks used by the AIO
tests, in combination with a TRY/CATCH block to reset the argument
value.  The infrastructure introduced in this commit will be reused for
the AIO tests, simplifying it.

Reviewed-by: Greg Burd <greg@burd.me>
Discussion: https://postgr.es/m/Z_y9TtnXubvYAApS@paquier.xyz