Fix reltuples stats after VACUUM FULL and CLUSTER
The tableam API's function for clustering a table, used by both
CLUSTER and VACUUM FULL, returns the number of non-removable tuples in
a num_tuples variable. This number includes dead tuples that cannot be
removed because they are still visible to concurrent
transactions. Previously, num_tuples was used to update
pg_class.reltuples, but this could lead to widely misleading numbers
if many tuples had been deleted or updated concurrently with ongoing
transactions. The fix is to not include the recently dead tuples.

It should be noted that this issue only occurred on tables that didn't
have any indexes, since the reindexing that happens as part of the
cluster code also updates reltuples, but using the more accurate
number based on currently visible tuples.