Add wait events statistics
Adding a new stat kind PGSTAT_KIND_WAIT_EVENT for the wait event statistics.

This new statistic kind is a fixed one because we know the maximum number of wait
events. Indeed:

 * it does not take into account custom wait events as extensions have all they need
 at their disposal to create custom stats on their own wait events should they
 want to (limited by WAIT_EVENT_CUSTOM_HASH_MAX_SIZE though).

 * it does not take into account LWLock > LWTRANCHE_FIRST_USER_DEFINED for the same
 reasons as above. That said, there is no maximum limitation in LWLockNewTrancheId().

 * we don't want to allocate memory in some places where the counters are
 incremented (see 4feba03d8b9). We could still use the same implementation as for
 backend statistics (i.e, make use of flush_static_cb) if we really want/need to
 switch to variable stats.

Some notes about the current design/implementation done in this patch:

For the moment only the counters are added (an array of currently 285 counters),
, weâll study/discuss about adding the timings once the counters are fully done.

I think weâd have more discussion/debate around the timings (should we add them
by default, add a new GUC, enable them at compilation time?â¦), thatâs why only
the counters are in this patch.

I think it makes sense as the counters have merit on their own.
We currently have 273 wait events but the array is 285 long: the reason is that
some wait events classes have "holes".

For some backends type the wait events stats are not flushed (walwriter for
example), so we need to find additional places to flush the wait events stats.

A few questions:

 * Do we need to serialize the stats based on their names (as for
 PGSTAT_KIND_REPLSLOT)? This question is the same as "is the ordering preserved
 if file stats format is not changed": I think the answer is yes (see f98dbdeb51d)
 , which means there is no need for to_serialized_name/from_serialized_name.

 * What if a new wait event is added? We'd need to change the stats file format,
 unless: the wait event stats kind becomes a variable one or we change a bit the
 way fixed stats are written/read to/from the stat file (we could add a new
 field in the PgStat_KindInfo for example).

XXX: Bump stat file format