Don't keep closed WAL segments in page cache after replay
On a standby, the recovery process reads the WAL segments, applies
changes and closes the segment. When closed, the segments will still be
in page cache memory until they are evicted due to inactivity. The
segments may be re-read if archive_mode is set to always, wal_summarizer
is enabled or if the standby is used for replication and has an active
walsender.

The presence of a replication slots is also a likely indicator that
a walsender will be started, and need to read the WAL segments.

Outside of those circumstances, the WAL segments won't be re-read and
keeping them in the page cache generates unnecessary memory pressure.
A POSIX_FADV_DONTNEED is sent before closing a replayed WAL segment to
immediately free any cached pages.