Fix documentation about CREATE INDEX CONCURRENTLY.
Fix the description of which transactions will block a
CREATE INDEX CONCURRENTLY command from proceeding.

This is actually a minor fixup for a65313f28bfc.

In DefineIndex(), we call WaitForLockers(heaplocktag, ShareLock, true),
and the heaplocktag is set with the relid of the heap, which means that
we wait for all the transactions that acquired a lock >= ROW EXCLUSIVE
on the heap.