Add new server config cleartext_passwords_action
This controls what happens when someone sends a clear text password
to the server via CREATE USER or ALTER USER. Three states are allowed:

1. "warn" The current default, this issues a warning if a clear
text password is used, but allows the change to proceed. The hint
changes to recommend \password if the current application_name is 'psql'

2. "allow" This does nothing, and thus emulates the historical behavior.

3. "disallow". This prevents the use of plain text completely, by throwing
an error if a password set or change is attempted.